pass color (#4f38a8) to tracedSVG in gatsby-remark-images config
https://www.gatsbyjs.org/packages/gatsby-remark-images/#options
tracedSVG
false
Use traced SVGs for placeholder images instead of the “blur up” effect.
Pass true for traced SVGs with the default settings (seen here
https://www.npmjs.com/package/gatsby-plugin-sharp#tracedsvg
), or an
object of options to override the defaults. For example, pass
{ color: "#F00", turnPolicy: "TURNPOLICY_MAJORITY" }
to change the color of the trace to red and the turn policy to
TURNPOLICY_MAJORITY. See node-potrace parameter documentation
https://github.com/tooolbox/node-potrace#parameters
for a full listing and explanation of the available options.